Upload center code cleaning
